+++ Owllery/Owllery/Extensions/UIViewExtension.swift
@@ -1,0 +1,20 @@
+//
+//  UIViewExtension.swift
+//  Owllery
+//
+//  Created by Nathalia Cardoso on 18/08/20.
+//  Copyright © 2020 Nathalia Cardoso. All rights reserved.
+//
+
+import UIKit
+
+extension UIView {
+    func pin(to superView: UIView) {
+        translatesAutoresizingMaskIntoConstraints = false
+        
+        topAnchor.constraint(equalTo: superView.topAnchor).isActive = true
+        leadingAnchor.constraint(equalTo: superView.leadingAnchor).isActive = true
+        trailingAnchor.constraint(equalTo: superView.trailingAnchor).isActive = true
+        bottomAnchor.constraint(equalTo: superView.bottomAnchor).isActive = true
+    }
+}
